feat: implemented navigation bar, header and content for the static version of the home page
Basic structure of the page has been added as well as styling, the content is divided into 2 sections, the style area which contains all of the CSS styling and the lower html area which contains all of the structure. the CSS area contains media queries to account for different screen sizes and adapt to them responsively.